BackEnd/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
googlemaps
requests
passlib[bcrypt]
python-jose[cryptography]
scikit-learn
numpy
scipy
bcrypt==4.0.1
